requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
APScheduler==3.10.4
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.7.3
pydantic_core==2.18.4
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.1
yadisk==2.1.0
yarl==1.9.4
aiogram
APScheduler
python-dotenv
